requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
crypto==1.4.1
cryptography==38.0.3
cycler==0.11.0
Deprecated==1.2.13
distlib==0.3.6
filelock==3.8.0
firebase-admin==6.0.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
gcloud==0.18.3
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.0.post0
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
msgpack==1.0.4
Naked==0.1.32
numpy==1.23.4
oauth2client==4.1.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
pipenv==2022.11.11
platformdirs==2.5.4
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.5.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.10.3
python-jwt==4.0.0
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.43
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.7
virtualenv-clone==0.5.7
Werkzeug==2.2.2
wrapt==1.14.1
